Recarga de pantalla y confirmacion de eliminacion de post

+++ src/app/my-post/my-post.component.ts
@@ -1,20 +1,22 @@
 import { Component, OnInit } from '@angular/core';
 import { Post } from '../model/post';
 import { PostService } from '../services/localApi/post.service';
-
+import { MatDialog, MatSnackBar } from '@angular/material';
+import { ConfirmarComponent } from '../dialogs/confirmar/confirmar.component';
 @Component({
   selector: 'app-my-post',
   templateUrl: './my-post.component.html',
   styleUrls: ['./my-post.component.scss']
 })
 export class MyPostComponent implements OnInit {
-
-  constructor(private postService: PostService) {
-  }
+  constructor(
+    private postService: PostService,
+    public dialog: MatDialog,
+    private snackBar: MatSnackBar
+  ) {}
   // tslint:disable-next-line:max-line-length
   postList: Post[];
   ngOnInit() {
-
     this.postService.getPostByUserLogin().subscribe(result => {
       this.postList = result;
       console.log(this.postList);
@@ -22,11 +24,24 @@ export class MyPostComponent implements OnInit {
   }
 
   eliminarEvento(idPost: number) {
-    // TO DO agregar dialog de confirmacion.
-    this.postService.deleteByPostId(idPost).subscribe(result => {
-      console.log(result);
-      // TO DO Recargar vista y mostrar tooltip de confirmacion.
+    const dialogRef = this.dialog.open(ConfirmarComponent, {
+      data: 'eliminarPost'
+    });
+    dialogRef.afterClosed().subscribe(response => {
+      if (response) {
+        this.postService.deleteByPostId(idPost).subscribe(result => {
+          console.log(result);
+          // this.postList.filter(p => p.idPost !== idPost);
+          this.ngOnInit();
+          this.snackBar.open(
+            'Se a eliminado el post exitosamente',
+            'aceptar',
+            {
+              duration: 2000
+            }
+          );
+        });
+      }
     });
   }
-
 }
